src/main/java/com/homework16/filters/AdminFilter.java
package com.homework16.filters;

import com.homework16.model.Role;
import java.io.IOException;
import javax.servlet.Filter;
import javax.servlet.FilterChain;
import javax.servlet.FilterConfig;
import javax.servlet.ServletException;
import javax.servlet.ServletRequest;
import javax.servlet.ServletResponse;
import javax.servlet.annotation.WebFilter;
import javax.servlet.http.HttpServletRequest;

/**
 * Filter for prevent go to admin page and do admin action.
 */
@WebFilter("/admin/*")
public class AdminFilter implements Filter {
  @Override
  public void init(FilterConfig filterConfig) throws ServletException {

  }

  @Override
  public void doFilter(ServletRequest servletRequest, ServletResponse servletResponse,
      FilterChain filterChain) throws IOException, ServletException {
    HttpServletRequest request = (HttpServletRequest) servletRequest;
    String role = (String) request.getSession().getAttribute("role");
    if (role == null || !role.equals(Role.ADMIN.getValue())) {
      request.getRequestDispatcher("/access_denied.jsp").forward(servletRequest, servletResponse);
    } else {
      filterChain.doFilter(servletRequest, servletResponse);
    }
  }

  @Override
  public void destroy() {

  }
}
